requirements.txt
flask==2.0.1
requests==2.26.0
python-dateutil==2.8.2
gunicorn==20.1.0 